% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IPS_coefs.R
\docType{data}
\name{IPS_coefs}
\alias{IPS_coefs}
\title{Initial coefficients of individual pension system data}
\format{
A list with 10 values, which are:
\describe{
\item{(Intercept)}{intercept}
\item{Labor_Rate}{labor rate}
\item{Unemployment_Rate}{unemployment rate}
\item{Sex_Ratio}{sex ratio}
\item{Urbanization_Rate}{urbanization rate}
\item{Deposit_Rate}{deposit rate}
\item{Illiteracy_Rate}{illiteracy rate}
\item{HDI}{human development index}
\item{lambda}{spatial autocorrelation parameter}
\item{s2}{variance}
}
}
\usage{
IPS_coefs
}
\description{
Initial coefficients of individual pension system data were obtained by MLE.
}
\keyword{datasets}
