% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IPS_data.R
\docType{data}
\name{IPS_data}
\alias{IPS_data}
\title{The individual pension system data of Turkey}
\format{
A list with 10 variables, which are:
\describe{
\item{ID}{provinces ID}
\item{Province}{provinces names}
\item{RPIPS}{participant rate of individual pension system}
\item{Labor_Rate}{labor rate}
\item{Unemployment_Rate}{unemployment rate}
\item{Sex_Ratio}{sex ratio}
\item{Urbanization_Rate}{urbanization rate}
\item{Deposit_Rate}{deposit rate}
\item{Illiteracy_Rate}{illiteracy rate}
\item{HDI}{human development index}
}
}
\usage{
IPS_data
}
\description{
This is individual pension system data of Turkey for analysing spatial error model.
}
\keyword{datasets}
