% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TRQWM.R
\docType{data}
\name{TRQWM}
\alias{TRQWM}
\title{Queen weight matrix of Turkey}
\format{
A symmetric matrix with 81x81 values,
\describe{
\item{V}{provinces ID}
}
}
\usage{
TRQWM
}
\description{
This is queen continugity weight matrix of Turkey.
}
\keyword{datasets}
