% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unemployment_coefs.R
\docType{data}
\name{unemployment_coefs}
\alias{unemployment_coefs}
\title{Initial coefficients of unemployment data}
\format{
A list with 4 values, which are:
\describe{
\item{(Intercept)}{intercept}
\item{Unemployment_Rate}{unemployment rate}
\item{lambda}{spatial autocorrelation parameter}
\item{s2}{variance}
}
}
\usage{
unemployment_coefs
}
\description{
Initial coefficients of unemployment data were obtained by MLE.
}
\keyword{datasets}
