% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unemployment_data.R
\docType{data}
\name{unemployment_data}
\alias{unemployment_data}
\title{Unemployment data of Turkey}
\format{
A list with 4 variables, which are:
\describe{
\item{ID}{provinces ID}
\item{province}{provinces names}
\item{unemployment}{unemployment rate}
\item{urbanization}{urbanization rate}
}
}
\usage{
unemployment_data
}
\description{
This is unemployment data of Turkey for analysing spatial error model.
}
\keyword{datasets}
