% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialRoMLE.R
\docType{package}
\name{SpatialRoMLE}
\alias{SpatialRoMLE}
\title{Spatial Robust MLE Package}
\description{
Robust Maximum Likelihood Estimation for Spatial Error Model.
}
\references{
Yildirim, V. and Kantar, Y.M. (2020). Robust estimation of spatial error model. in Journal of Statistical Computation and Simulation. \url{https://doi.org/10.1080/00949655.2020.1740223}
}
\author{
Vural Yildirim \email{vurall_yildirim@hotmail.com}

Yeliz Mert Kantar
}
